cleanup cluster_config and slurm status
cluster_config deprecated in favor of profiles
the standalone sequana_slurm_status is removed.
pipeline should use the manager.error_report  function instead
