+++ sequana_pipetools/snaketools/pipeline_utils.py
@@ -61,7 +61,6 @@ class OnSuccessCleaner:
         self.files_to_remove = [
             "config.yaml",
             "multiqc_config.yaml",
-            "cluster_config.json",
             "slurm*out",
             "stats.txt",
             "unlock.sh",
